requirements/docs.txt
-e .
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --extra-index-url=https://pkgs.frkl.io/frkl/dev -o requirements/docs.txt requirements/docs.in
#
--extra-index-url https://pkgs.frkl.io/frkl/dev
asn1crypto==0.24.0        # via cryptography
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
blessings==1.7
cerberus==1.2
certifi==2019.3.9         # via requests
cffi==1.12.3              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via click-completion, click-log
click-completion==0.5.0
click-log==0.3.2
colorama==0.3.9           # via halo, log-symbols
coverage==4.5.3           # via pytest-cov
cryptography==2.6.1       # via pyopenssl, requests
cursor==1.2.0             # via halo
decorator==4.4.0          # via dogpile.cache
dogpile.cache==0.7.1
filelock==3.0.10          # via tox
future==0.17.1
halo==0.0.24
idna==2.8                 # via requests
jinja2==2.10.1            # via click-completion
log-symbols==0.0.14       # via halo
markdown==3.1             # via mdv
markupsafe==1.1.1         # via jinja2
mdv==1.7.4
more-itertools==7.0.0     # via pytest
passlib==1.7.1
pbr==5.2.0                # via stevedore
pluggy==0.9.0             # via pytest, tox
plumbum==1.6.7
pp-ez==0.2.0
pprintpp==0.4.0
py==1.8.0                 # via pytest, tox
pycparser==2.19           # via cffi
pygments==2.3.1           # via mdv
pyopenssl==19.0.0         # via requests
pytest==4.4.1             # via pytest-cov
pytest-cov==2.6.1
requests[security]==2.21.0
ruamel.yaml.jinja2==0.2.2  # via ruamel.yaml
ruamel.yaml[jinja2]==0.15.94  # via ruamel.yaml.jinja2
shellingham==1.3.1        # via click-completion
six==1.12.0               # via blessings, click-completion, cryptography, halo, pyopenssl, pytest, stevedore, tox
spinners==0.0.23          # via halo
stevedore==1.30.1
tabulate==0.8.3           # via mdv
termcolor==1.1.0          # via halo
toml==0.10.0              # via tox
tox==3.9.0
tqdm==4.31.1
treelib==1.5.5
urllib3==1.24.2           # via requests
virtualenv==16.5.0        # via tox
